requirements.txt
fastapi
pyyaml
uvicorn
requests